Move p-memo implementation
